MLIBZ-2366: Rename before-prepare hook to before-checkForChanges
