+++ packages/kinvey-nativescript-sdk/platforms/android/AndroidManifest.default.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
+	<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
 	<application android:name="com.tns.NativeScriptApplication">
 		<activity android:name="com.tns.NativeScriptActivity">
 			<intent-filter>
